Refactor mock PCS server into separate module
